requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
boto3
python-multipart
celery[redis]
redis
python-dotenv
httpx
pytest
requests